refactor(rust): errors: fix ockam_vault
