+++ implementations/rust/ockam/ockam_vault/src/error.rs
@@ -1,59 +1,83 @@
-use ockam_core::Error;
+use ockam_core::{
+    error::{
+        code::{ErrorCode, Kind, Origin},
+        Error2,
+    },
+    thiserror,
+};
 
 /// Represents the failures that can occur in
 /// an Ockam vault
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, thiserror::Error)]
 pub enum VaultError {
     /// Secret does not belong to this vault
+    #[error("secret does not belong to this vault")]
     SecretFromAnotherVault = 1,
     /// Public key is invalid
+    #[error("public key is invalid")]
     InvalidPublicKey,
     /// Unknown ECDH key type
+    #[error("unknown ECDH key type")]
     UnknownEcdhKeyType,
     /// Invalid key type
+    #[error("invalid key type")]
     InvalidKeyType,
     /// Entry not found
+    #[error("entry not found")]
     EntryNotFound,
     /// Invalid AES key length
+    #[error("invalid AES key length")]
     InvalidAesKeyLength,
     /// Invalid Secret length
     InvalidSecretLength,
     /// Invalid HKDF output type
     InvalidHkdfOutputType,
     /// Invalid private key length
+    #[error("invalid private key length")]
     InvalidPrivateKeyLen,
     /// AES encryption failed
+    #[error("aes encryption failed")]
     AeadAesGcmEncrypt,
     /// AES decryption failed
+    #[error("aes decryption failed")]
     AeadAesGcmDecrypt,
     /// HKDF key expansion failed
+    #[error("hkdf key expansion failed")]
     HkdfExpandError,
     /// Secret not found
+    #[error("secret not found")]
     SecretNotFound,
     /// Invalid X25519 secret length
+    #[error("invalid X25519 secret length")]
     InvalidX25519SecretLength,
     /// Invalid Ed25519 secret
+    #[error("invalid Ed25519 secret")]
     InvalidEd25519Secret,
     /// Invalid BLS secret length
+    #[error("invalid BLS secret length")]
     InvalidBlsSecretLength,
     /// Invalid BLS secret
+    #[error("invalid BLS secret")]
     InvalidBlsSecret,
     /// IO error when saving
     StorageError,
 }
 
-impl VaultError {
-    /// Integer code associated with the error domain.
-    pub const DOMAIN_CODE: u32 = 12_000;
-    /// Descriptive name for the error domain.
-    pub const DOMAIN_NAME: &'static str = "OCKAM_VAULT";
-}
-
-impl From<VaultError> for Error {
+impl From<VaultError> for Error2 {
     fn from(err: VaultError) -> Self {
-        Self::new(
-            VaultError::DOMAIN_CODE + (err as u32),
-            ockam_core::compat::format!("{}::{:?}", module_path!(), err),
-        )
+        use VaultError::*;
+        let kind = match err {
+            SecretFromAnotherVault
+            | InvalidPublicKey
+            | InvalidKeyType
+            | InvalidAesKeyLength
+            | InvalidHkdfOutputType
+            | InvalidPrivateKeyLen
+            | InvalidX25519SecretLength => Kind::Misuse,
+            UnknownEcdhKeyType | EntryNotFound | SecretNotFound => Kind::NotFound,
+            _ => Kind::Invalid,
+        };
+
+        Error2::new(ErrorCode::new(Origin::Vault, kind), err)
     }
 }
